Create Parking-Dilemma.py
There are many cars parked in a parking lot. The parking lot is a straight line with a parking spot for every meter. There are n cars currently parked and a roofer wants to cover them with a roof. The requirement is that at least k cars currently in the lot are covered by the roof. Determined the minimum length of the roof. Determine the minimum length of the roof to cover k cars.

Example:
n = 
cars = [6, 2, 12, 7]k = 3