Competitive Gaming
A group of friends are playing a video game together. During the game, each player earns a number of points. At the end of a round, players who achieve at least a certain rank get to "Level up" their character to gain increased abilities. Given the scores of the players at the end of a round, how many players will be able to level up?