Disk Space Analysis
A company is performing an analysis on the computer at its main office. The computer are spaced along a single row. For each group of the contiguous computers of a certain length, that is, for each segment, determine the minimum amount of disk space available on a computer. Return the maximum of these values as your answer. 

Example:
n = 4, the number of computers
space = [8, 2, 4, 6]
x = 2, the segment length

Function Description....


Source:
https://github.com/kashifusmani/interview_prep/blob/bce8e1393509f032cdb5b6ab907d4c52b06f6c66/ssense/questions.py
https://github.com/cbenavidesdev/algorithms/blob/8086d9bdee6ee9f326225fc65f006b50bbe33dd3/OA/disk_space_analysis.py
https://github.com/JoshuaSamuelTheCoder/Quarantine/blob/9cad9115418c44b2d83e091956347abb57d03294/Disk_Space_Analysis/main.py